requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
cryptography==3.0
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.8
gunicorn==20.0.4
httplib2==0.18.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
pycparser==2.20
pylint==2.5.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
